Added Creation Directory when uploading
The system will now automatically create a folder whenever a user upload
a file. Unlike before it cause an error because there is no existing
folder when the upload could be save. The folder created is based on
user_id so it will not be easily replaced because it is unique.

Things need to be add soon
 - Friendly Message when user uploaded a picture with maximum size.

 - Error trapping for files.

+ How does this change address the issue?
+ What side effects does this change have?