Added Form Validatio in Registration and Upload
By having the form validation it prevents the unnecessary error that
may arrive in the future upon registration. Also in Uploading new item
validation already created to make sure user input is correct

How does it address the issue? For short obvious patches this part can be omitted, but it should be a high level description of what the approach was.

Add
	- Email validation when uploaded online