Refactor Code in Registartion Controller
Split the process of one function that all process is being integrated.
The registration function seems to do all necessary validation and
saving and checking to the database. It is moved to Model_User to
handle all of the database functionality.

Model_User now handling all of the saving and checking of user that
try to register

By doing those changes , it helps the future developer to easily add
functionaily .