Finished Messaging System with Scrolling Functionality
User can now navigate into old conversation by just using the scroll.
It enables the user to free some spaces but see the old conversation.

The scrollbar triggers a command into AJAX that will request into
server if there is still a messages left. It uses the concept of
creating a page number like 1,2,3,4 by with automatick without the
need of clicking the button